requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-socketio==5.10.0
python-engineio==4.8.0
Werkzeug==3.0.1
eventlet==0.33.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
